+++ test/ueber_facebook_test.exs
@@ -1,6 +1,6 @@
-defmodule UeberauthFacebookTest do
+defmodule UeberauthInstagramTest do
   use ExUnit.Case
-  doctest UeberauthFacebook
+  doctest UeberauthInstagram
 
   test "the truth" do
     assert 1 + 1 == 2